+++ src/js/content/cellar/tiles/base.js
@@ -1,22 +1,45 @@
 content.cellar.tiles.base = {
-  effects : undefined,
-  note: undefined,
-  prime: undefined,
+  id: undefined,
+  alwaysAudible: false,
+  isUnique: false,
+  name: undefined,
+  synthType: 'sine',
   uniquePerFloor: false,
   uniquePerRun: false,
   weight: 1,
+  // Instance properties
+  effects : undefined,
+  note: undefined,
+  prime: undefined,
+  seed: undefined,
+  state: undefined,
   x: undefined,
   y: undefined,
   z: undefined,
+  // State
+  defaultState: {},
   // Generator
   generate: function (tile = {}) {
     return this.extend(tile)
   },
   // Methods
   extend: function (definition = {}) {
-    return engine.fn.extend(this, definition)
+    const instance = engine.fn.extend(this, definition)
+
+    instance.defaultState = {...this.defaultState, ...instance.defaultState}
+    instance.state = {...instance.defaultState}
+
+    return instance
+  },
+  getName: function () {
+    return this.name
   },
   getEffects: function () {
     return this.effects || []
   },
+  isFullyScanned: function () {
+    return content.cellar.scans.get(this) == this.getEffects().length
+  },
+  onEnter: function () {},
+  onExit: function () {},
 }